Revert "make message compatible between systems"
This reverts commit 8ae913b2e757ba9fc5919f350b6e84bc6f5c8987.

+++ communicate/include/message.hpp
@@ -4,10 +4,7 @@
 #include <string>
 #include <memory.h>
 
-// This is the maximum number of characters in a message.
-// note that it it should never exceed 256.
 #define MAX_MSG_SIZE 200
-
 class Message
 {
 public:
@@ -16,7 +13,7 @@ public:
     std::string get_msg();
 
 private:
-    uint8_t length;
+    size_t length;
     char buf[MAX_MSG_SIZE];
 };
 #endif
